Add deadline to offer schema and create offers table
